requirements.txt
asgiref==3.8.1
Django==5.1.7
filelock==3.18.0
fsspec==2025.3.0
gunicorn==23.0.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4

packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
setuptools==78.1.0
sqlparse==0.5.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0


typing_extensions==4.13.0


